requirements.txt
asgiref==3.4.1
Django==3.2.8
pandas==2.1.4
powerline-status==2.7
PyMySQL==1.0.2
pytz==2021.3
sqlparse==0.4.2
django-simpleui==2023.12.12
PyJWT==2.8.0
Pillow==10.1.0
